#2093 fixed ConcurrentModificationException; updated to last version of ATSD HTTP API; v.0.4.1
git-svn-id: https://localhost/svn/axibase/AxibaseAPI/atsd-api-java/trunk@11599 c89e5cec-f22d-a84c-b7a1-b0d050e0f212

+++ src/main/java/com/axibase/tsd/example/AtsdClientWriteExample.java
@@ -91,8 +91,8 @@ public class AtsdClientWriteExample extends AbstractAtsdClientExample {
             }
             memoryEater.add(sb.toString());
 
-            double totalMemoryMb = runtime.totalMemory() / MB;
-            double freeMemoryMb = runtime.freeMemory() / MB;
+            long totalMemoryMb = runtime.totalMemory() / MB;
+            long freeMemoryMb = runtime.freeMemory() / MB;
             sendToAtsd(totalMemoryMb, freeMemoryMb);
 
             try {
